+++ packages/shorebird_cli/lib/src/commands/patch/patch_aar_command.dart
@@ -177,7 +177,11 @@ Please re-run the release command for this version or create a new release.''');
     return await runScoped(
       () async {
         final buildNumber = results['build-number'] as String;
-        final buildProgress = logger.progress('Building patch');
+        final flutterVersionString =
+            await shorebirdFlutter.getVersionAndRevision();
+        final buildProgress = logger.progress(
+          'Building patch with Flutter $flutterVersionString',
+        );
         try {
           await buildAar(buildNumber: buildNumber);
           buildProgress.complete();