+++ packages/shorebird_cli/lib/src/commands/patch/patch_android_command.dart
@@ -139,7 +139,11 @@ If this option is not provided, the version number will be determined from the p
       releaseVersion = argReleaseVersion;
     } else {
       logger.detail('No release version provided. Determining from bundle.');
-      final buildProgress = logger.progress('Building patch');
+      final flutterVersionString =
+          await shorebirdFlutter.getVersionAndRevision();
+      final buildProgress = logger.progress(
+        'Building patch with Flutter $flutterVersionString',
+      );
       try {
         await buildAppBundle(flavor: flavor, target: target);
         buildProgress.complete();