+++ packages/shorebird_cli/test/src/shorebird_env_test.dart
@@ -236,6 +236,27 @@ void main() {
           isA<Pubspec>().having((p) => p.name, 'name', 'test'),
         );
       });
+
+      test(
+          'returns value when pubspec.yaml exists '
+          'and contains a malformed value', () {
+        final tempDir = Directory.systemTemp.createTempSync();
+        File(
+          p.join(tempDir.path, 'pubspec.yaml'),
+        ).writeAsStringSync('''
+name: test
+publish_to: yon30c
+        ''');
+        expect(
+          IOOverrides.runZoned(
+            () => runWithOverrides(() => shorebirdEnv.getPubspecYaml()),
+            getCurrentDirectory: () => tempDir,
+          ),
+          isA<Pubspec>()
+              .having((p) => p.name, 'name', 'test')
+              .having((p) => p.publishTo, 'publishTo', isNull),
+        );
+      });
     });
 
     group('hasPubspecYaml', () {
@@ -255,6 +276,23 @@ void main() {
         File(
           p.join(tempDir.path, 'pubspec.yaml'),
         ).writeAsStringSync('name: test');
+        expect(
+          IOOverrides.runZoned(
+            () => runWithOverrides(() => shorebirdEnv.hasPubspecYaml),
+            getCurrentDirectory: () => tempDir,
+          ),
+          isTrue,
+        );
+      });
+
+      test('returns true even if pubspec.yaml contains malformed values', () {
+        final tempDir = Directory.systemTemp.createTempSync();
+        File(
+          p.join(tempDir.path, 'pubspec.yaml'),
+        ).writeAsStringSync('''
+name: test
+publish_to: yon30c
+        ''');
         expect(
           IOOverrides.runZoned(
             () => runWithOverrides(() => shorebirdEnv.hasPubspecYaml),