+++ packages/shorebird_code_push_protocol/lib/src/messages/create_patch/create_patch_request.g.dart
@@ -16,31 +16,17 @@ CreatePatchRequest _$CreatePatchRequestFromJson(Map<String, dynamic> json) =>
         final val = CreatePatchRequest(
           releaseId: $checkedConvert('release_id', (v) => v as int),
           wasForced: $checkedConvert('was_forced', (v) => v as bool?),
-          hasAssetChanges:
-              $checkedConvert('has_asset_changes', (v) => v as bool?),
-          hasNativeChanges:
-              $checkedConvert('has_native_changes', (v) => v as bool?),
-          metadata: $checkedConvert(
-              'metadata',
-              (v) => v == null
-                  ? null
-                  : CreatePatchMetadata.fromJson(v as Map<String, dynamic>)),
+          metadata: $checkedConvert('metadata',
+              (v) => CreatePatchMetadata.fromJson(v as Map<String, dynamic>)),
         );
         return val;
       },
-      fieldKeyMap: const {
-        'releaseId': 'release_id',
-        'wasForced': 'was_forced',
-        'hasAssetChanges': 'has_asset_changes',
-        'hasNativeChanges': 'has_native_changes'
-      },
+      fieldKeyMap: const {'releaseId': 'release_id', 'wasForced': 'was_forced'},
     );
 
 Map<String, dynamic> _$CreatePatchRequestToJson(CreatePatchRequest instance) =>
     <String, dynamic>{
       'release_id': instance.releaseId,
       'was_forced': instance.wasForced,
-      'has_asset_changes': instance.hasAssetChanges,
-      'has_native_changes': instance.hasNativeChanges,
-      'metadata': instance.metadata?.toJson(),
+      'metadata': instance.metadata.toJson(),
     };