+++ RELEASE_NOTES.md
@@ -2,19 +2,27 @@
 
 This section contains past updates we've sent to customers via Discord.
 
-## 1.0.1 (April 16, 2024)
+## 1.0.1 (April 17, 2024)
 
 - 🚀 Improve link percentage by ~10-15% for all iOS patches
 - 🩹 Fix potentially incorrect behavior in optimized try/catch on iOS patches
 - 🥞 Fix potential crashes when using dwarf stack traces on iOS patches
 - 🖇️ Fix potential errors on iOS patches due to link table alignment bug in the snapshot
-- 🔐 Ensure all release artifacts are built using the same Flutter revision
+- 💥 Fix issue where patching a release with multiple platforms built with different flutter versions would crash.
 - 🪵 Improve logs to only include stack traces when using `--verbose`
+- 🔍 Fix issue where `shorebird init` would not find a `pubspec.yaml` due to malformed values
 - ❅ Fix flaky failures when encoding archives (observed in `shorebird preview` on Windows)
 - 📝 Include patch instructions after all `release` commands
 - 🔗 Link to the Shorebird Console when a release already exists
 - 🚉 Support `--target-platforms` on android release and patch commands
 - 🍧 Fix `shorebird release android` with camelCase flavor names
+
+This is a required upgrade due to the issue present in all previous `shorebird`
+versions where it was possible to create invalid (crashing) patches if the same
+release number was used for multiple platforms but built with different Flutter
+versions.
+
+Please upgrade using `shorebird upgrade`.
 
 ℹ️ [Shorebird Status](https://docs.shorebird.dev/status) tracks any known issues.